Version API spec to V1, refactor schema into files (#493)
This should make the API specs much easier to maintain, especially as we
create new API versions. The entire API spec exists under spec/api/vX;
openapi schemas are separated into files depending on resource; utility
functions related to API schemas are separated into their own util
class. Note that this is just a refactoring and does not change the
openapi.json (other than some example ID for some reason).
