+++ spec/api/v1/schemas/rule_results.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+module Api
+  module V1
+    module Schemas
+      module RuleResults
+        RULE_RESULT = {
+          type: 'object',
+          required: %w[result],
+          properties: {
+            result: {
+              type: 'string',
+              example: 'passed'
+            }
+          }
+        }.freeze
+      end
+    end
+  end
+end
