+++ swagger/v1/openapi.json
@@ -415,7 +415,7 @@
                   "relationships": {
                     "account": {
                       "data": {
-                        "id": "cf8ae1be-2626-4c7b-8134-bda1c78946c6",
+                        "id": "6a888174-88fd-4e80-a100-25550e5f77e6",
                         "type": "account"
                       }
                     },
